fix(validation): add handler for idempotencyToken
The `idempotencyToken` flag is set on many different API parameter
definitions. It has no real implication for validation because API
definitions that use it tend to include additional string validation
operators for it, such as min/max length and regex match.

I suspect that this field is present as a hint to the SDK that it should
generate a UUID if the caller has not specified one, because here's an
excerpt from the docs for one of the APIs that accepts such a parameter:

> If you use the AWS CLI or one of the AWS SDKs to call this operation,
> then you can leave this parameter empty because they generate a random
> UUID for you.

See also:
  * https://docs.aws.amazon.com/secretsmanager/latest/apireference/API_PutSecretValue.html#SecretsManager-PutSecretValue-request-ClientRequestToken