fix(*): fix AWS_CONTAINER_CREDENTIALS_FULL_URI parsing (#65)
The RemoteCredential does not support a full URI with port number now, the error log shows as follow:

```
2023/07/27 15:48:41 [debug] 46876#0: [lua] RemoteCredentials.lua:70: Failed to construct RemoteCredentials url: Unsupported hostname: AWS.RemoteCredentials only supports localhost,127.0.0.1 for http; http://localhost requested.
```

It seems that the code is using wrong reference for host, it should be host instead of hostname.

The document of socket.url.parse:

```
url.parse(url, default)

Parses an URL given as a string into a Lua table with its components.

Url is the URL to be parsed. If the default table is present, it is used to store the parsed fields. Only fields present in the URL are overwritten. Therefore, this table can be used to pass default values for each field.

The function returns a table with all the URL components:

parsed_url = {
  url = string,
  scheme = string,
  authority = string,
  path = string,
  params = string,
  query = string,
  fragment = string,
  userinfo = string,
  host = string,
  port = string,
  user = string,
  password = string
}
```